build: refactor manpage generation to be a quasi-test
