+++ qcp/src/cli/mod.rs
@@ -3,5 +3,5 @@
 mod args;
 mod cli_main;
 pub mod styles;
-pub use args::CliArgs;
 pub use cli_main::cli;
+mod manpage;
