+++ xtask/Cargo.toml
@@ -8,9 +8,5 @@ publish = false
 
 [dependencies]
 anyhow = { workspace = true }
-clap = { workspace = true }
-clap_mangen = "0.2.26"
 pico-args = "0.5.0"
-qcp = { path = "../qcp" }
-roff = "0.2.2"
 xshell = "0.2.7"
